Fix compilation errors - Add missing model properties and fix type issues
Fixes:
1. Added VisualImagePaths property to VideoRequest model
2. Added OutputSettings property to VideoRequest model
3. Added VisualCue computed property to ScriptSegment model
4. Fixed Math.Ceiling ambiguous call by explicit double cast
5. Fixed AnimateDiff array type errors (mixed string/int to object[])

All 13 compilation errors resolved.

+++ src/Services/VideoGenerationPipeline.cs
@@ -114,7 +114,7 @@ public class VideoGenerationPipeline
             var prompt = new VideoPrompt
             {
                 Description = segment.VisualCue,
-                Duration = (int)Math.Ceiling(segment.Duration),
+                Duration = (int)Math.Ceiling((double)segment.Duration),
                 AspectRatio = _aiVideoConfig.DefaultSettings.AspectRatio,
                 Style = _aiVideoConfig.DefaultSettings.Style,
                 MotionIntensity = _aiVideoConfig.DefaultSettings.MotionIntensity,
